ETL/2_orchestrator/requirements.txt
functions-framework==3.*
google-cloud-logging
requests
